backend/requirements.txt
Flask
supabase
openai
python-dotenv
flask_cors
langchain
langchain_groq